requirements.txt
pefile
python-daemon
python-magic
ssdeep
